Graphics.Gnuplot.ColorSpecification

Documentation

data T

rgb :: Double -> Double -> Double -> T

Color components for Red, Green, Blue, must be in the range [0,1].

rgb8 :: Word8 -> Word8 -> Word8 -> T

name :: String -> T

Specify a color by name. You are responsible for choosing an existing name in gnuplot. If your color is available as Haskell variable in this module, then prefer this one.

paletteFrac :: Double -> T

white :: T

black :: T

gray0 :: T

grey0 :: T

gray10 :: T

grey10 :: T

gray20 :: T

grey20 :: T

gray30 :: T

grey30 :: T

gray40 :: T

grey40 :: T

gray50 :: T

grey50 :: T

gray60 :: T

grey60 :: T

gray70 :: T

grey70 :: T

gray80 :: T

grey80 :: T

gray90 :: T

grey90 :: T

gray100 :: T

grey100 :: T

gray :: T

grey :: T

lightGray :: T

lightGrey :: T

darkGray :: T

darkGrey :: T

red :: T

lightRed :: T

darkRed :: T

yellow :: T

lightYellow :: T

darkYellow :: T

green :: T

lightGreen :: T

darkGreen :: T

springGreen :: T

forestGreen :: T

seaGreen :: T

blue :: T

lightBlue :: T

darkBlue :: T

midnightBlue :: T

navy :: T

mediumBlue :: T

royalblue :: T

skyblue :: T

cyan :: T

lightCyan :: T

darkCyan :: T

magenta :: T

lightMagenta :: T

darkMagenta :: T

turquoise :: T

lightTurquoise :: T

darkTurquoise :: T

pink :: T

lightPink :: T

darkPink :: T

coral :: T

lightCoral :: T

orangeRed :: T

salmon :: T

lightSalmon :: T

darkSalmon :: T

aquamarine :: T

khaki :: T

darkKhaki :: T

goldenrod :: T

lightGoldenrod :: T

darkGoldenrod :: T

gold :: T

beige :: T

brown :: T

orange :: T

darkOrange :: T

violet :: T

darkViolet :: T

plum :: T

purple :: T