Graphics.Gnuplot.Encoding

Description

Support for special characters

Gnuplot has no universal Unicode escaping mechanism, you can only work with encodings. However, not all terminals support all encodings, not all terminals even support utf-8. Some terminals seem to support only one encoding. E.g. WX seems to support only UTF-8, X11 seems to support only Latin-1. Postscript, SVG, PNG seem to support both UTF-8 and Latin-1.

The gnuplot Haskell bindings always write using the system-wide default encoding. Thus it is better not to set an encoding other than locale explicitly. However, if you write the files yourself in a certain encoding you should use the encoding option of the according terminal.

Documentation

data T

locale :: T

deflt :: T

iso_8859_1 :: T

iso_8859_15 :: T

iso_8859_2 :: T

iso_8859_9 :: T

koi8r :: T

koi8u :: T

cp437 :: T

cp850 :: T

cp852 :: T

cp950 :: T

cp1250 :: T

cp1251 :: T

cp1254 :: T

sjis :: T

utf8 :: T