Graphics.Gnuplot.Frame.Option

Documentation

data T

Every option represents an internal state in gnuplot. It is altered with gnuplot's set command. The first field in T is the name of the option and the name of the according internal state in gnuplot.

Sometimes the addressed state is not explicitly mentioned but is expressed by the syntax of the values. E.g. you can write set grid xtics and set grid noxtics, but both commands refer to the same internal boolean variable, that we like to call xtics. It is important that the gnuplot Haskell bindings know that these two set commands refer to the same gnuplot state, since we want to simulate a stateless functional interface in front of a stateful imperative one.

In case of a such a hidden state, we manage an identifier in the second field of T. It is mainly used for distinguishing different hidden states, that are accessed by the same set variable. This second field may not contain valid gnuplot identifiers, however you might use the field for formatting boolean options using addBool.

custom :: String -> String -> T

Constructs a generic option from Strings for the first and second field of T.

This is very flexible, but not very safe. Use it only as fall-back, if there is no specific setter function in Graphics.Gnuplot.Frame.OptionSet.

title :: T

grid :: String -> T

size :: String -> T

key :: String -> T

border :: String -> T

pm3d :: String -> T

view :: T

xRange :: String -> T

yRange :: String -> T

zRange :: String -> T

xLabel :: String -> T

yLabel :: String -> T

zLabel :: String -> T

xTicks :: String -> T

yTicks :: String -> T

zTicks :: String -> T

xLogScale :: T

yLogScale :: T

zLogScale :: T

xData :: T

yData :: T

zData :: T

xFormat :: T

yFormat :: T

zFormat :: T

timeFmt :: T