Graphics.Gnuplot.Frame.OptionSet

Documentation

data T graph

deflt :: C graph => T graph

add :: T -> [String] -> T graph -> T graph

Add (set) an option with arguments as plain strings.

This is very flexible, but not very safe. Use it only as fall-back, if there is no specific setter function in Graphics.Gnuplot.Frame.OptionSet.

remove :: T -> T graph -> T graph

Remove (unset) an option.

This is very flexible, but not very safe. Use it only as fall-back, if there is no specific setter function in Graphics.Gnuplot.Frame.OptionSet.

boolean :: T -> Bool -> T graph -> T graph

Set or unset option according to a Bool. This is for switches that can be disabled using unset.

This is very flexible, but not very safe. Use it only as fall-back, if there is no specific setter function in Graphics.Gnuplot.Frame.OptionSet.

See also: addBool, add, remove.

addBool :: T -> Bool -> T graph -> T graph

Add an option with boolean value that is formatted like set style fill border and set style fill noborder. The name of the internal state (i.e. border) must be stored in the second field of the option.

This is very flexible, but not very safe. Use it only as fall-back, if there is no specific setter function in Graphics.Gnuplot.Frame.OptionSet.

See also boolean.

size :: C graph => Double -> Double -> T graph -> T graph

sizeRatio :: C graph => Double -> T graph -> T graph

sizeSquare :: C graph => T graph -> T graph

title :: C graph => String -> T graph -> T graph

key :: C graph => Bool -> T graph -> T graph

keyInside :: C graph => T graph -> T graph

keyOutside :: C graph => T graph -> T graph

xRange2d :: (C x, C y, C x) => (x, x) -> T (T x y) -> T (T x y)

yRange2d :: (C x, C y, C y) => (y, y) -> T (T x y) -> T (T x y)

xRange3d :: (C x, C y, C z, C x) => (x, x) -> T (T x y z) -> T (T x y z)

yRange3d :: (C x, C y, C z, C y) => (y, y) -> T (T x y z) -> T (T x y z)

zRange3d :: (C x, C y, C z, C z) => (z, z) -> T (T x y z) -> T (T x y z)

xLabel :: C graph => String -> T graph -> T graph

yLabel :: C graph => String -> T graph -> T graph

zLabel :: (C x, C y, C z) => String -> T (T x y z) -> T (T x y z)

xTicks2d :: (C x, C y, C x) => [(String, x)] -> T (T x y) -> T (T x y)

yTicks2d :: (C x, C y, C y) => [(String, y)] -> T (T x y) -> T (T x y)

xTicks3d :: (C x, C y, C z, C x) => [(String, x)] -> T (T x y z) -> T (T x y z)

yTicks3d :: (C x, C y, C z, C y) => [(String, y)] -> T (T x y z) -> T (T x y z)

zTicks3d :: (C x, C y, C z, C z) => [(String, z)] -> T (T x y z) -> T (T x y z)

xLogScale :: C graph => T graph -> T graph

yLogScale :: C graph => T graph -> T graph

zLogScale :: (C x, C y, C z) => T (T x y z) -> T (T x y z)

grid :: C graph => Bool -> T graph -> T graph

gridXTicks :: C graph => Bool -> T graph -> T graph

gridYTicks :: C graph => Bool -> T graph -> T graph

gridZTicks :: (C x, C y, C z) => Bool -> T (T x y z) -> T (T x y z)

xFormat :: C graph => String -> T graph -> T graph

yFormat :: C graph => String -> T graph -> T graph

zFormat :: (C x, C y, C z) => String -> T (T x y z) -> T (T x y z)

view

Arguments

:: Double
rotateX

-> Double
rotateZ

-> Double
scale

-> Double
scaleZ

-> T (T x y z)

-> T (T x y z)

Set parameters of viewing a surface graph. See info:gnuplot/view

viewMap :: T (T x y z) -> T (T x y z)

Show flat pixel map.

boxwidthRelative :: C graph => Double -> T graph -> T graph

boxwidthAbsolute :: C graph => Double -> T graph -> T graph