Graphics.Gnuplot.Plot.ThreeDimensional

Documentation

type T x y z = T (T x y z)

Plots can be assembled using mappend or mconcat or several functions from Data.Foldable.

cloud :: (C x, C y, C z, C a) => T x y z a -> [a] -> T x y z

mesh :: (C x, C y, C z, C x, C y, C z) => [[(x, y, z)]] -> T x y z

surface :: (C x, C y, C z, C x, C y, C z) => [x] -> [y] -> (x -> y -> z) -> T x y z

linearScale :: Fractional a => Integer -> (a, a) -> [a]

functionToGraph :: [x] -> (x -> y) -> [(x, y)]