Graphics.Gnuplot.Terminal.PNG

Documentation

data T

cons :: FilePath -> T

encoding :: T -> T -> T

Setting the encoding to anything different from locale makes only sense if you write your gnuplot files manually using this encoding.

transparent :: T -> T

noTransparent :: T -> T

interlace :: T -> T

noInterlace :: T -> T

trueColor :: T -> T

noTrueColor :: T -> T

fontTiny :: T -> T

fontSmall :: T -> T

fontMedium :: T -> T

fontLarge :: T -> T

fontGiant :: T -> T