Graphics.Gnuplot.Terminal.PostScript

Documentation

data T

cons :: FilePath -> T

encoding :: T -> T -> T

Setting the encoding to anything different from locale makes only sense if you write your gnuplot files manually using this encoding.

landscape :: T -> T

portrait :: T -> T

eps :: T -> T

color :: T -> T

monochrome :: T -> T

font :: String -> Int -> T -> T

embedFont :: FilePath -> T -> T

Embed a font file in the generated PostScript output. Each call adds a new font file, there is no way to remove it again.