Graphics.Gnuplot.Terminal.SVG

Documentation

data T

cons :: FilePath -> T

encoding :: T -> T -> T

Setting the encoding to anything different from locale makes only sense if you write your gnuplot files manually using this encoding.