Graphics.Gnuplot.Value.Atom

Description

Provide a class that restricts the range of Haskell types to the ones that gnuplot can process.

Documentation

data OptionSet a

Constructors

OptionSet

Fields

optData :: [String]

optFormat :: [String]

optOthers :: [(T, [String])]

class C a where

Minimal complete definition

Nothing

Methods

options :: OptionSet a