Graphics.Gnuplot.Value.ColumnSet

Description

We provide a way to specify a set of columns that matches the tuple structure of a certain graph type.

Documentation

newtype T a

Constructors

Cons [Int]

atom :: C a => Int -> T a

pair :: T a -> T b -> T (a, b)

triple :: T a -> T b -> T c -> T (a, b, c)

quadruple :: T a -> T b -> T c -> T d -> T (a, b, c, d)