Graphics.Gnuplot.Value.Tuple

Description

Provide a class that renders multiple Haskell values in a text form that is accessible by gnuplot.

Maybe we add a method for the binary interface to gnuplot later.

Documentation

class C a where

Minimal complete definition

text

Methods

text :: a -> [ShowS]

For values that are also in Atom class, text must generate a singleton list.

columnCount :: ColumnCount a

It must hold ColumnCount (length (text x)) == columnCount.

newtype ColumnCount a

Count numbers of gnuplot data columns for the respective type.

Somehow a writer monad with respect to Sum monoid without material monadic result.

Cf. ColumnSet module.

Constructors

ColumnCount Int

newtype Label

Constructors

Label String